+++ .github/templates/attack/rollback/gl0_cluster/action.yml
@@ -1,17 +1,22 @@
-name: Rollback Global L0 Cluster
+name: Double Signing Global L0 Cluster
+
+inputs:
+  RUN_ROLLBACK:
+    required: false
 
 runs:
   using: "composite"
   steps:
-    - name: Run Rollback L0 Genesis
+    - name: Run Double Signing L0 Genesis
       with:
         CL_PUBLIC_HTTP_PORT: 9000
         CL_P2P_HTTP_PORT: 9001
         CL_CLI_HTTP_PORT: 9002
-      uses: "./.github/templates/attack/rollback_l0/genesis"
+        RUN_ROLLBACK: ${{ inputs.RUN_ROLLBACK }}
+      uses: "./.github/templates/attack/rollback/genesis"
 
-    - name: Run Rollback L0 Validator - 1
-      uses: "./.github/templates/attack/rollback_l0/validator"
+    - name: Run Double Signing L0 Validator - 1
+      uses: "./.github/templates/attack/rollback/validator"
       with:
         CL_PUBLIC_HTTP_PORT: 8000
         CL_P2P_HTTP_PORT: 8001
@@ -20,8 +25,8 @@ runs:
         CL_L0_P2P_HTTP_PORT: 9001
         NODE_NUMBER: 1
 
-    - name: Run Rollback L0 Validator - 2
-      uses: "./.github/templates/attack/rollback_l0/validator"
+    - name: Run Double Signing L0 Validator - 2
+      uses: "./.github/templates/attack/rollback/validator"
       with:
         CL_PUBLIC_HTTP_PORT: 8100
         CL_P2P_HTTP_PORT: 8101
