+++ .github/templates/project_template/modules/doubleSign/src/main/scala/org/tessellation/dag/l0/attack/doubleSign/Main.scala
@@ -1,13 +1,12 @@
 package org.tessellation.dag.l0.inspect.attack.doubleSign
 
-import cats.effect.Async
+import cats.effect.IO
 
 import org.tessellation.dag.l0.DagL0Application
 import org.tessellation.dag.l0.infrastructure.snapshot.GlobalSnapshotArtifact
-import org.tessellation.kryo.KryoSerializer
 import org.tessellation.node.shared.domain.gossip.Gossip
 import org.tessellation.node.shared.infrastructure.snapshot.GossipForkInfo
-import org.tessellation.security.{Hasher, SecurityProvider}
+import org.tessellation.security.Hasher
 
 import eu.timepit.refined.auto._
 
@@ -17,7 +16,7 @@ object Main extends DagL0Application {
 
   override def mkGossipForkInfo(
     gossip: Gossip[IO]
-  )(implicit SP: SecurityProvider[IO], K: KryoSerializer[IO], H: Hasher[IO]): GossipForkInfo[IO, GlobalSnapshotArtifact] =
+  )(implicit H: Hasher[IO]): GossipForkInfo[IO, GlobalSnapshotArtifact] =
     GossipDoubleSign.make(gossip, doubleSignConfig.offset)
 
 }
