+++ .github/templates/project_template/project/Dependencies.scala
@@ -3,10 +3,11 @@ import sbt._
 object Dependencies {
 
   object V {
-    //Default version of tessellation develop branch
+    // Default version of tessellation develop branch
     val tessellation = "99.99.99"
     val decline = "2.4.1"
   }
+
   def tessellation(artifact: String): ModuleID = "org.constellation" %% s"tessellation-$artifact" % V.tessellation
 
   def decline(artifact: String = ""): ModuleID =
@@ -15,6 +16,7 @@ object Dependencies {
     } % V.decline
   object Libraries {
     val tessellationKernel = tessellation("kernel")
+    val tessellationDAGL0 = tessellation("dag-l0")
     val tessellationDAGL1 = tessellation("dag-l1")
     val tessellationNodeShared = tessellation("node-shared")
     val tessellationShared = tessellation("shared")
@@ -26,7 +28,6 @@ object Dependencies {
     val declineEffect = decline("effect")
     val declineRefined = decline("refined")
   }
-
 
   // Scalafix rules
   val organizeImports = "com.github.liancheng" %% "organize-imports" % "0.5.0"
