+++ .github/workflows/detect_rollback.yml
@@ -1,0 +1,150 @@
+name: Test rollback is not double signing
+
+on:
+  pull_request: { }
+  push:
+    branches:
+      - develop
+      - hotfix/*
+      - release/*
+
+jobs:
+  build:
+    name: Test rollback is not double signing
+    runs-on: ubuntu-20.04
+    timeout-minutes: 45
+    strategy:
+      matrix:
+        java:
+          - openjdk@1.11.0
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: actions/setup-node@v3
+        with:
+          node-version: 16.14.2
+          cache: 'npm'
+          cache-dependency-path: .github/action_scripts/package-lock.json
+
+      - name: Installing node-fetch@2.6.6
+        run: |
+          cd .github/action_scripts
+          npm i node-fetch@2.6.6
+
+      - name: Setup Java and scala
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          JAVA_VERSION: ${{ matrix.java }}
+        uses: "./.github/templates/setup_java_and_scala"
+
+      - name: Update version.sbt to github-action-test-version
+        shell: bash
+        run: |
+          echo 'ThisBuild / version := "99.99.99"' > version.sbt
+
+      - name: Generating Metagraph dependencies
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          sbt --error clean compile shared/publishM2 dagL0/publishM2 kernel/publishM2 keytool/publishM2 nodeShared/publishM2 currencyL0/publishM2 currencyL1/publishM2 dagL1/publishM2
+
+      - name: Run Global L0 Cluster
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        uses: "./.github/templates/global_l0/gl0_cluster"
+
+      - name: Run Rollback L0 Cluster
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          RUN_ROLLBACK: true
+        uses: "./.github/templates/attack/rollback/gl0_cluster"
+
+      - name: Run Metagraph L1 Cluster
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        uses: "./.github/templates/metagraph_l1/ml1_cluster"
+
+      - name: Check Metagraph cluster
+        run: |
+          cd .github/action_scripts/check_clusters
+          node metagraph.js
+
+      - name: Check rewards
+        run: |
+          cd .github/action_scripts/rewards
+          node index.js
+
+      - name: Wait 1 minute before check cluster
+        run: |
+          sleep 60
+
+      ## Checking cluster again because we already faced scenarios where the cluster dies after sending rewards transactions
+      - name: Check Metagraph cluster
+        run: |
+          cd .github/action_scripts/check_clusters
+          node metagraph.js
+
+      - name: Save Global L0 Genesis logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-global-l0-genesis-java-${{ matrix.java }}
+          path: .github/config/containers/global-l0/genesis/global-l0-genesis.log
+
+      - name: Save Global L0 - Validator 1 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-global-l0-validator-1-java-${{ matrix.java }}
+          path: .github/config/containers/global-l0/validator-1/global-l0-validator-1.log
+
+      - name: Save Global L0 - Validator 2 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-global-l0-validator-2-java-${{ matrix.java }}
+          path: .github/config/containers/global-l0/validator-2/global-l0-validator-2.log
+
+      - name: Save Metagraph L0 - Genesis logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l0-genesis-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l0/genesis/metagraph-l0-genesis.log
+
+      - name: Save Metagraph L0 - Validator 1 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l0-validator-1-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l0/validator-1/metagraph-l0-validator-1.log
+
+      - name: Save Metagraph L0 - Validator 2 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l0-validator-2-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l0/validator-2/metagraph-l0-validator-2.log
+
+      - name: Save Metagraph L1 - Initial Validator logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l1-initial-validator-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l1/initial-validator/metagraph-l1-initial-validator.log
+
+      - name: Save Metagraph L1 - Validator 1 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l1-validator-1-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l1/validator-1/metagraph-l1-1.log
+
+      - name: Save Metagraph L1 - Validator 2 logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: rewards-metagraph-l1-validator-2-java-${{ matrix.java }}
+          path: .github/config/containers/metagraph-l1/validator-2/metagraph-l1-2.log
